header now client component so it updates correctly
